# Optimize reduce