Various improvements to the reduce function with the goal of improving speed and reducing memory usage.
The main architectural changes are:
* Replace the huge numpy color mapping array by a dictionary data structure
* Pre-compute lab palette instead of re-running the conversion every time nearest_color_ciede2000 is called

The benchmarks presented below were run on [the following image](https://static01.nyt.com/images/2021/09/14/science/07CAT-STRIPES/07CAT-STRIPES-mediumSquareAt3X-v2.jpg), using `timeit` and `memit` for profiling:
|               | New     | Old         |
|---------------|---------|-------------|
| fast-time     | 209 ms  | 220 ms      |
| fast-mem      | 241 MiB | 319.43 MiB  |
| accurate-time | 765 ms  | 2230 ms     |
| accurate-mem  | 270 MiB | 1262.61 MiB |

 NB: memory is still being leaked for the accurate method, albeit at a much slower rate than before.