@Seon82 2022-03-05
Now also fixes the memory leak in `rgb2lab`