@GrayTurtles 2022-03-05
This is indeed much better for both the memory and the speed and thanks for fixing the memory leak.
However you removed some optimizations I had in the old version so I think this could be improved even more:

1. Populate the initial `cache` dictionary with the palette: this makes the function way faster when it's used on an image that already matches the palette (which often happens) because there's no computation to do.
2. Why iterate over every pixel? Wouldn't it be faster to iterate over the unique colors like I did and then apply the `cache` color map to the initial array as bytes?
3. I set `cached=True` in some `@jit` decorators so the function get cached and they don't have to get re-compiled after each reboot but I couldn't do it for the CIEDE2000 functions because I was using `parallel=True` and it doesn't work with the cache. Since you removed the parallel, maybe the cache could be added on the missing functions.