@Seon82 2022-03-05
1. `nearest_color_idx_ciede2000` currently runs in ~6.75 µs. This means that even if all the palette colors are used by a template, we'll only be gaining 0.2 ms out of several hundreds. Imo it's not worth it.
3. Iterating over each pixel directly has the advantage of reducing the total number of iterations (we're iterating over each pixel in image vs in image+unique color array), and saves a bit of memory. Applying a dict map to an array is also a lot slower than the coordinate lookup you introduced, so I strongly suspect this will end up being slower (haven't tested tho).
4. Will do!